+++ CHANGELOG.md
@@ -17,7 +17,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * New field type, "date"
 * New field type, "datetime"
-* Static site: Show size of SQLite database
+* Static site: 
+  * Show size of SQLite database
+  * Errors page shows all errors, site and record
 
 ## [0.3.1] - 2022-03-31